ssh plugin PAM

+++ python/uninstall.py
@@ -27,19 +27,54 @@ import os
 import shutil
 
 from latchHelper import *
+from translation import *
 
 
-# read sshd_config file
-f = open(SSHD_CONFIG,"r");
-lines = f.readlines();
-f.close();
-# delete latch
-f = open(SSHD_CONFIG,"w");
-for line in lines:
-    if line.find("ForceCommand " + WRAPPER_EXE) == -1 :
-        f.write(line);
-f.close();
 
+if os.path.isfile(SSHD_PAM_CONFIG_FILE):
+    # read PAM config file
+    f = open(SSHD_PAM_CONFIG_FILE,"r");
+    lines = f.readlines();
+    f.close();
+
+    shutil.copy(SSHD_PAM_CONFIG_FILE, SSHD_PAM_CONFIG_FILE + "~")
+
+    # delete latch PAM 
+    f = open(SSHD_PAM_CONFIG_FILE,"w");
+    for line in lines:
+        if equalSplit(line, AUTH_INCLUDE_PASSWD_AUTH_SSHD):
+            # centos version
+            f.write(AUTH_INCLUDE_PASSWORD_AUTH + "\n")
+        elif not equalSplit(line, AUTH_REQUIRED_LATCH_PAM):
+            # ubuntu version
+            f.write(line)
+    f.close()
+else:
+    print(CANT_OPEN_SSHD_PAM_MSG)
+
+
+if os.path.isfile(SSHD_CONFIG):
+    # read sshd_config file
+    f = open(SSHD_CONFIG,"r")
+    lines = f.readlines()
+    f.close()
+
+    shutil.copy(SSHD_CONFIG, SSHD_CONFIG + "~")
+
+    # delete latch
+    f = open(SSHD_CONFIG,"w")
+    for line in lines:
+        if "ChallengeResponseAuthentication" in line and "#" not in line:
+            f.write("ChallengeResponseAuthentication no\n")
+        elif "PasswordAuthentication" in line and "#" not in line:
+            f.write("PasswordAuthentication yes\n")
+        elif not equalSplit(line, "ForceCommand " + WRAPPER_EXE):
+            f.write(line);
+    f.close();
+else:
+    print(CANT_OPEN_SSHD_CONFIG_MSG)
+
+# uninstall files
 if os.path.isfile(LATCH_ACCOUNTS):
     os.remove(LATCH_ACCOUNTS)
 if os.path.isfile(LATCH_CONFIG):
@@ -50,6 +85,8 @@ if os.path.isfile(WRAPPER_PY):
     os.remove(WRAPPER_PY)
 if os.path.isfile(WRAPPER_EXE):
     os.remove(WRAPPER_EXE)
+if os.path.isfile(PASSWORD_AUTH_SSHD_PAM_CONFIG_FILE):
+    os.remove(PASSWORD_AUTH_SSHD_PAM_CONFIG_FILE)
 if os.path.isdir(LATCH_OPENSSH_PATH):
     shutil.rmtree(LATCH_OPENSSH_PATH)
 if os.path.isfile(PAIR_BIN):
@@ -61,4 +98,4 @@ if os.path.isfile(PLUGIN_BIN):
 if os.path.isfile(SETTINGS_BIN):
     os.remove(SETTINGS_BIN)
 
-print("Uninstall completed")
+print(UNINSTALLED_MSG)
